+++ v2/examples/example_client_roundrobin_test.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -31,8 +31,8 @@ import (
 // ExampleNewRoundRobinEndpoints shows how to create a client with round-robin endpoint list
 func ExampleNewRoundRobinEndpoints() {
 	// Create an HTTP connection to the database
-	endpoints := connection.NewRoundRobinEndpoints([]string{"https://a:8529", "https://a:8539", "https://b:8529"})
-	conn := connection.NewHttpConnection(exampleJSONHTTPConnectionConfig(endpoints))
+	endpoint := connection.NewRoundRobinEndpoints([]string{"https://a:8529", "https://a:8539", "https://b:8529"})
+	conn := connection.NewHttp2Connection(connection.DefaultHTTP2ConfigurationWrapper(endpoint, true))
 
 	// Create a client
 	client := arangodb.NewClient(conn)